put Tree class into its own file
